move db save to another file
